+++ client/CHANGELOG.md
@@ -2,7 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
-## v1.48.0 - Unreleased
+## v1.49.0 - Unreleased
+
+### Added
+
+- Added the timeline on place pages ([#517](https://github.com/project-lux/lux-frontend/issues/517)).
+- Added Collections Created At, Collections About, and Collection Types to Place pages ([#623](https://github.com/project-lux/lux-frontend/issues/623)).
+- Added alt text to images for accessibility purposes ([#616](https://github.com/project-lux/lux-frontend/issues/616)).
+- Added images to concept pages and results snippets ([#602](https://github.com/project-lux/lux-frontend/issues/602)).
+
+### Changed
+
+- Changed CMS content page rendering ([#645](https://github.com/project-lux/lux-frontend/issues/645)).
+
+### Removed
+
+### Fixed
+
+## v1.48.0 - 2025-4-7
 
 ### Added